Align IRnode source with LUA receiver, Default PIR/LDR to enabled in roomNode2
git-svn-id: https://livebox-hah.googlecode.com/svn/trunk@281 64644fc6-01da-11df-bca2-d32778fd5087

+++ userapps/arduino/IRNode/IRNode.pde
@@ -20,8 +20,8 @@ static byte myNodeID;       // node ID used for this unit
 
 // This defines the structure of the packets which get sent out by wireless:
 struct {
-  int type; // NEC, SONY, RC5, UNKNOWN
-  int bits; // Number of bits in decoded value
+  byte type; // NEC, SONY, RC5, UNKNOWN
+  byte bits; // Number of bits in decoded value
   unsigned long value;  // decoded value
 } 
 payload;
@@ -81,5 +81,4 @@ void loop () {
     irrecv.resume(); // Receive the next value
   }
 }
-
